requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
pydantic>=2.0.0
PyYAML>=6.0.0
